+++ main.c
@@ -6,9 +6,9 @@
 int main(int argc, char **argv)
 {
 	int *xs = NULL;
-	int size = 0;
-	int len = 0;
-	int i;
+	size_t size = 0;
+	size_t len = 0;
+	size_t i;
 	if (read_array(&xs, &size, &len) < 0) {
 		if (errno == ENOMEM) {
 			fprintf(stderr, "out of memory\n");
